+++ lib/s_offstage.dart
@@ -3,4 +3,4 @@
 /// This library exports all public APIs for the s_offstage package.
 library;
 
-export 'src/my_offstage.dart';
+export 'src/s_offstage.dart';